Refactored all code in model_funcs.py over to model_history.py.
