+++ torchlens/model_history.py
@@ -22,7 +22,6 @@ from torchlens.helper_funcs import get_attr_values_from_tensor_list, get_tensor_
     log_current_rng_states, make_random_barcode, make_short_barcode_from_input, make_var_iterable, \
     move_input_tensors_to_device, print_override, remove_entry_from_list, safe_copy, set_random_seed, \
     set_rng_from_saved_states, tuple_tolerant_assign
-from torchlens.model_funcs import cleanup_model, prepare_model
 from torchlens.torch_decorate import decorate_pytorch, undecorate_pytorch
 
 
@@ -794,6 +793,304 @@ class ModelHistory:
 
         return model_df
 
+    ###########################
+    ##### Model Functions #####
+    ###########################
+
+    def prepare_model(self,
+                      model: nn.Module,
+                      module_orig_forward_funcs: Dict,
+                      decorated_func_mapper: Dict[Callable, Callable]):
+        """Adds annotations and hooks to the model, and decorates any functions in the model.
+
+        Args:
+            model: Model to prepare.
+            module_orig_forward_funcs: Dict with the original forward funcs for each submodule
+            decorated_func_mapper: Dictionary mapping decorated functions to original functions, so they can be restored
+
+        Returns:
+            Model with hooks and attributes added.
+        """
+        self.model_name = str(type(model).__name__)
+        model.tl_module_address = ''
+        model.tl_source_model_history = self
+
+        module_stack = [(model, '')]  # list of tuples (name, module)
+
+        while len(module_stack) > 0:
+            module, parent_address = module_stack.pop()
+            module_children = list(module.named_children())
+
+            # Decorate any torch functions in the model:
+            for func_name, func in module.__dict__.items():
+                if (func_name[0:2] == '__') or (not callable(func)) or (func not in decorated_func_mapper):
+                    continue
+                module.__dict__[func_name] = decorated_func_mapper[func]
+
+            # Annotate the children with the full address.
+            for c, (child_name, child_module) in enumerate(module_children):
+                child_address = f"{parent_address}.{child_name}" if parent_address != '' else child_name
+                child_module.tl_module_address = child_address
+                module_children[c] = (child_module, child_address)
+            module_stack = module_children + module_stack
+
+            if module == model:  # don't tag the model itself.
+                continue
+
+            module.tl_source_model_history = self
+            module.tl_module_type = str(type(module).__name__)
+            self.module_types[module.tl_module_address] = module.tl_module_type
+            module.tl_module_pass_num = 0
+            module.tl_module_pass_labels = []
+            module.tl_tensors_entered_labels = []
+            module.tl_tensors_exited_labels = []
+
+            # Add decorators.
+
+            if hasattr(module, 'forward'):
+                module_orig_forward_funcs[module] = module.forward
+                module.forward = module_forward_decorator(module.forward, module, model_history)
+                module.forward.tl_forward_call_is_decorated = True
+
+        # Mark all parameters with requires_grad = True, and mark what they were before, so they can be restored on cleanup.
+        for param in model.parameters():
+            param.tl_requires_grad = param.requires_grad
+            param.requires_grad = True
+
+        # And prepare any buffer tensors.
+        self.prepare_buffer_tensors(model)
+
+    def prepare_buffer_tensors(self,
+                               model: nn.Module):
+        """Goes through a model and all its submodules, and prepares any "buffer" tensors: tensors
+        attached to the module that aren't model parameters.
+
+        Args:
+            model: PyTorch model
+
+        Returns:
+            PyTorch model with all buffer tensors prepared and ready to track.
+        """
+        submodules = self.get_all_submodules(model)
+        for submodule in submodules:
+            for attribute_name in dir(submodule):
+                attribute = getattr(submodule, attribute_name)
+                if issubclass(type(attribute), torch.Tensor) and not issubclass(type(attribute), torch.nn.Parameter):
+                    if submodule.tl_module_address == '':
+                        buffer_address = attribute_name
+                    else:
+                        buffer_address = submodule.tl_module_address + '.' + attribute_name
+                    self.log_source_tensor(attribute, 'buffer', buffer_address)
+
+    def module_forward_decorator(self,
+                                 orig_forward: Callable,
+                                 module: nn.Module) -> Callable:
+        @wraps(orig_forward)
+        def decorated_forward(*args, **kwargs):
+
+            # "Pre-hook" operations:
+            module_address = module.tl_module_address
+            module.tl_module_pass_num += 1
+            module_pass_label = (module_address, module.tl_module_pass_num)
+            module.tl_module_pass_labels.append(module_pass_label)
+            input_tensors = get_vars_of_type_from_obj([args, kwargs], torch.Tensor)
+            for t in input_tensors:
+                tensor_entry = self[t.tl_tensor_label_raw]
+                module.tl_tensors_entered_labels.append(t.tl_tensor_label_raw)
+                tensor_entry.modules_entered.append(module_address)
+                tensor_entry.module_passes_entered.append(module_pass_label)
+                tensor_entry.is_submodule_input = True
+                tensor_entry.module_entry_exit_thread_output.append(('+', module_pass_label[0], module_pass_label[1]))
+
+            # The function call
+            out = orig_forward(*args, **kwargs)
+
+            # "Post-hook" operations:
+            module_address = module.tl_module_address
+            module_pass_num = module.tl_module_pass_num
+            module_entry_label = module.tl_module_pass_labels.pop()
+            output_tensors = get_vars_of_type_from_obj(out, torch.Tensor)
+            for t in output_tensors:
+                if module.tl_module_type.lower() == 'identity':  # if identity module, run the function for bookkeeping
+                    t = getattr(torch, 'identity')(t)
+                tensor_entry = self[t.tl_tensor_label_raw]
+                tensor_entry.is_submodule_output = True
+                tensor_entry.is_bottom_level_submodule_output = self.log_whether_exited_submodule_is_bottom_level(t,
+                                                                                                                  module)
+                tensor_entry.modules_exited.append(module_address)
+                tensor_entry.module_passes_exited.append((module_address, module_pass_num))
+                tensor_entry.module_entry_exit_thread_output.append(('-', module_entry_label[0], module_entry_label[1]))
+                module.tl_tensors_exited_labels.append(t.tl_tensor_label_raw)
+
+            for t in input_tensors:  # Now that module is finished, roll back the threads of all input tensors.
+                tensor_entry = self[t.tl_tensor_label_raw]
+                input_module_thread = tensor_entry.module_entry_exit_thread_output[:]
+                if ('+', module_entry_label[0], module_entry_label[1]) in input_module_thread:
+                    module_entry_ix = input_module_thread.index(('+', module_entry_label[0], module_entry_label[1]))
+                    tensor_entry.module_entry_exit_thread_output = tensor_entry.module_entry_exit_thread_output[
+                                                                   :module_entry_ix]
+
+            return out
+
+        return decorated_forward
+
+    def log_whether_exited_submodule_is_bottom_level(self,
+                                                     t: torch.Tensor,
+                                                     submodule: nn.Module):
+        """Checks whether the submodule that a tensor is leaving is a "bottom-level" submodule;
+        that is, that only one tensor operation happened inside the submodule.
+
+        Args:
+            t: the tensor leaving the module
+            submodule: the module that the tensor is leaving
+
+        Returns:
+            Whether the tensor operation is bottom level.
+        """
+        tensor_entry = self[getattr(t, 'tl_tensor_label_raw')]
+        submodule_address = submodule.tl_module_address
+
+        if tensor_entry.is_bottom_level_submodule_output:
+            return True
+
+        # If it was initialized inside the model and nothing entered the module, it's bottom-level.
+        if tensor_entry.initialized_inside_model and len(submodule.tl_tensors_entered_labels) == 0:
+            tensor_entry.is_bottom_level_submodule_output = True
+            tensor_entry.bottom_level_submodule_pass_exited = (submodule_address, submodule.tl_module_pass_num)
+            return True
+
+        # Else, all parents must have entered the submodule for it to be a bottom-level submodule.
+        for parent_label in tensor_entry.parent_layers:
+            parent_tensor = self[parent_label]
+            parent_modules_entered = parent_tensor.modules_entered
+            if (len(parent_modules_entered) == 0) or (parent_modules_entered[-1] != submodule_address):
+                tensor_entry.is_bottom_level_submodule_output = False
+                return False
+
+        # If it survived the above tests, it's a bottom-level submodule.
+        tensor_entry.is_bottom_level_submodule_output = True
+        tensor_entry.bottom_level_submodule_pass_exited = (submodule_address, submodule.tl_module_pass_num)
+        return True
+
+    def get_all_submodules(self,
+                           model: nn.Module,
+                           is_top_level_model: bool = True) -> List[nn.Module]:
+        """Recursively gets list of all submodules for given module, no matter their level in the
+        hierarchy; this includes the model itself.
+
+        Args:
+            model: PyTorch model.
+            is_top_level_model: Whether it's the top-level model; just for the recursive logic of it.
+
+        Returns:
+            List of all submodules.
+        """
+        submodules = []
+        if is_top_level_model:
+            submodules.append(model)
+        for module in model.children():
+            submodules.append(module)
+            submodules += self.get_all_submodules(module, is_top_level_model=False)
+        return submodules
+
+    def cleanup_model(self,
+                      model: nn.Module,
+                      module_orig_forward_funcs: Dict[nn.Module, Callable],
+                      model_device: str,
+                      decorated_func_mapper: Dict[Callable, Callable]):
+        """Reverses all temporary changes to the model (namely, the forward hooks and added
+        model attributes) that were added for PyTorch x-ray (scout's honor; leave no trace).
+
+        Args:
+            model: PyTorch model.
+            model_device: Device the model is stored on
+            module_orig_forward_funcs: Dict containing the original, undecorated forward pass functions for each submodule
+            decorated_func_mapper: Dict mapping between original and decorated PyTorch funcs
+
+        Returns:
+            Original version of the model.
+        """
+        submodules = self.get_all_submodules(model, is_top_level_model=True)
+        for submodule in submodules:
+            if submodule == model:
+                continue
+            submodule.forward = module_orig_forward_funcs[submodule]
+        self.restore_model_attributes(model, decorated_func_mapper=decorated_func_mapper, attribute_keyword='tl')
+        self.undecorate_model_tensors(model, model_device)
+
+    def clear_hooks(self,
+                    hook_handles: List):
+        """Takes in a list of tuples (module, hook_handle), and clears the hook at that
+        handle for each module.
+
+        Args:
+            hook_handles: List of tuples (module, hook_handle)
+
+        Returns:
+            Nothing.
+        """
+        for hook_handle in hook_handles:
+            hook_handle.remove()
+
+    def restore_module_attributes(self,
+                                  module: nn.Module,
+                                  decorated_func_mapper: Dict[Callable, Callable],
+                                  attribute_keyword: str = 'tl'):
+        for attribute_name in dir(module):
+            if attribute_name.startswith(attribute_keyword):
+                delattr(module, attribute_name)
+                continue
+            attr = getattr(module, attribute_name)
+            if isinstance(attr, Callable) and (attr in decorated_func_mapper) and (attribute_name[0:2] != '__'):
+                setattr(module, attribute_name, decorated_func_mapper[attr])
+
+    def restore_model_attributes(self,
+                                 model: nn.Module,
+                                 decorated_func_mapper: Dict[Callable, Callable],
+                                 attribute_keyword: str = 'tl'):
+        """Recursively clears the given attribute from all modules in the model.
+
+        Args:
+            model: PyTorch model.
+            decorated_func_mapper: Dict mapping between original and decorated PyTorch funcs
+            attribute_keyword: Any attribute with this keyword will be cleared.
+
+        Returns:
+            Nothing.
+        """
+        for module in self.get_all_submodules(model):
+            self.restore_module_attributes(module,
+                                           decorated_func_mapper=decorated_func_mapper,
+                                           attribute_keyword=attribute_keyword)
+
+        for param in model.parameters():
+            param.requires_grad = getattr(param, 'tl_requires_grad')
+            for attribute_name in dir(param):
+                if attribute_name.startswith(attribute_keyword):
+                    delattr(param, attribute_name)
+
+    def undecorate_model_tensors(model: nn.Module, model_device: str):
+        """Goes through a model and all its submodules, and unmutates any tensor attributes. Normally just clearing
+        parameters would have done this, but some module types (e.g., batchnorm) contain attributes that are tensors,
+        but not parameters.
+
+        Args:
+            model: PyTorch model
+            model_device: device the model is stored on
+
+        Returns:
+            PyTorch model with unmutated versions of all tensor attributes.
+        """
+        submodules = get_all_submodules(model)
+        for submodule in submodules:
+            for attribute_name in dir(submodule):
+                attribute = getattr(submodule, attribute_name)
+                if issubclass(type(attribute), torch.Tensor):
+                    if not issubclass(type(attribute), torch.nn.Parameter):
+                        setattr(submodule, attribute_name, undecorate_tensor(attribute, model_device))
+                    else:
+                        remove_attributes_starting_with_str(attribute, 'tl_')
+
     def get_op_nums_from_user_labels(self, which_layers: List[Union[str, int]]) -> List[int]:
         """Given list of user layer labels, returns the original tensor numbers for those labels (i.e.,
         the numbers that were generated on the fly during the forward pass, such that they can be
@@ -851,9 +1148,9 @@ class ModelHistory:
         Returns:
             Nothing, but now the ModelHistory object will have saved activations for the new input.
         """
-        raise NotImplementedError  # todo: implement this
-        '''
         self.logging_mode = 'fast'
+
+        # Get the tensor numbers to save based on user-specified layers.
         if layers_to_save != 'all':
             tensor_nums_to_save = []
             bad_layers = []
@@ -867,9 +1164,28 @@ class ModelHistory:
                 raise ValueError(f"Following desired layers not logged in ModelHistory: \n\t{bad_layers}")
         else:
             tensor_nums_to_save = 'all'
-    
-        self.tensor_nums_to_save = tensor_nums_to_save
-        '''
+
+        # Go through and clear all existing activations.
+        for tensor_log_entry in self:
+            tensor_log_entry.tensor_contents = None
+            tensor_log_entry.has_saved_activations = False
+            tensor_log_entry.has_saved_grad = False
+            tensor_log_entry.grad_contents = None
+
+        # Reset relevant fields.
+        self.layers_with_saved_activations = []
+        self.layers_with_saved_gradients = []
+        self.num_tensors_saved = 0
+        self.tensor_fsize_saved = 0
+        self.tensor_counter = 0
+        self.raw_layer_type_counter = defaultdict(lambda: 0)
+
+        # Now run and log the new inputs.
+        self.run_and_log_inputs_through_model(model,
+                                              input_args,
+                                              input_kwargs,
+                                              tensor_nums_to_save,
+                                              random_seed)
 
     ########################################
     #### Running and logging the model #####
@@ -1179,7 +1495,6 @@ class ModelHistory:
             self.layers_with_saved_activations.append(orig_tensor_entry.layer_label)
             orig_tensor_entry.save_tensor_data(t, [], {}, self.save_function_args, self.activation_postfunc)
 
-        orig_tensor_entry
         orig_tensor_entry.tensor_shape = tuple(t.shape)
         orig_tensor_entry.tensor_dtype = t.dtype
         orig_tensor_entry.tensor_fsize = get_tensor_memory_amount(t)
@@ -1460,7 +1775,7 @@ class ModelHistory:
             orig_tensor_entry.func_rng_states = func_rng_states
             orig_tensor_entry.func_position_args_non_tensor = non_tensor_args
             orig_tensor_entry.func_keyword_args_non_tensor = non_tensor_kwargs
-        
+
     @staticmethod
     def _output_should_be_logged(out: Any,
                                  is_bottom_level_func: bool) -> bool:
@@ -1990,6 +2305,10 @@ class ModelHistory:
         """
         After the forward pass, cleans up the log into its final form.
         """
+        if self.logging_mode == 'fast':
+            self.postprocess_fast(decorated_func_mapper)
+            return
+
         # Step 1: Add dedicated output nodes
 
         self._add_output_layers()
@@ -2042,6 +2361,13 @@ class ModelHistory:
 
         # Step 14: log the pass as finished, changing the ModelHistory behavior to its user-facing version.
 
+        self._set_pass_finished()
+
+    def postprocess_fast(self, decorated_func_mapper: Dict):
+        self._trim_and_reorder_model_history_fields()
+        self._remove_unwanted_entries_and_log_remaining()
+        self._undecorate_all_saved_tensors(decorated_func_mapper)
+        torch.cuda.empty_cache()
         self._set_pass_finished()
 
     def _add_output_layers(self):
@@ -2968,6 +3294,13 @@ class ModelHistory:
                 self.num_tensors_saved += 1
 
         i = 0
+        self.layer_list = []
+        self.layer_dict_main_keys = {}
+        self.layer_labels = []
+        self.layer_labels_no_pass = []
+        self.layer_labels_w_pass = []
+        self.layer_num_passes = {}
+
         for tensor_entry in self:
             # Determine valid lookup keys and relate them to the tensor's realtime operation number:
             if tensor_entry.has_saved_activations or self.keep_unsaved_layers:
@@ -4549,3 +4882,55 @@ class ModelHistory:
 
     def __repr__(self):
         return self.__str__()
+
+
+def run_model_and_save_specified_activations(model: nn.Module,
+                                             input_args: Union[torch.Tensor, List[Any]],
+                                             input_kwargs: Dict[Any, Any],
+                                             tensor_nums_to_save: Optional[Union[str, List[int]]] = 'all',
+                                             keep_unsaved_layers: bool = True,
+                                             output_device: str = 'same',
+                                             activation_postfunc: Optional[Callable] = None,
+                                             mark_input_output_distances: bool = False,
+                                             detach_saved_tensors: bool = False,
+                                             save_function_args: bool = False,
+                                             save_gradients: bool = False,
+                                             random_seed: Optional[int] = None) -> ModelHistory:
+    """Internal function that runs the given input through the given model, and saves the
+    specified activations, as given by the tensor numbers (these will not be visible to the user;
+    they will be generated from the nicer human-readable names and then fed in).
+
+    Args:
+        model: PyTorch model.
+        input_args: Input arguments to the model's forward pass: either a single tensor, or a list of arguments.
+        input_kwargs: Keyword arguments to the model's forward pass.
+        tensor_nums_to_save: List of tensor numbers to save.
+        keep_unsaved_layers: Whether to keep layers in the ModelHistory log if they don't have saved activations.
+        output_device: device where saved tensors will be stored: either 'same' to keep unchanged, or
+            'cpu' or 'cuda' to move to cpu or cuda.
+        activation_postfunc: Function to apply to activations before saving them (e.g., any averaging)
+        mark_input_output_distances: Whether to compute the distance of each layer from the input or output.
+            This is computationally expensive for large networks, so it is off by default.
+        detach_saved_tensors: whether to detach the saved tensors, so they remain attached to the computational graph
+        save_function_args: whether to save the arguments to each function
+        save_gradients: whether to save gradients from any subsequent backward pass
+        random_seed: Which random seed to use.
+
+    Returns:
+        ModelHistory object with full log of the forward pass
+    """
+    model_name = str(type(model).__name__)
+    model_history = ModelHistory(model_name,
+                                 output_device,
+                                 activation_postfunc,
+                                 keep_unsaved_layers,
+                                 save_function_args,
+                                 save_gradients,
+                                 detach_saved_tensors,
+                                 mark_input_output_distances)
+    model_history.run_and_log_inputs_through_model(model,
+                                                   input_args,
+                                                   input_kwargs,
+                                                   tensor_nums_to_save,
+                                                   random_seed)
+    return model_history
